requirements.txt
oss2>=2.19.1,<3.0.0
aliyun-python-sdk-core>=2.16.0,<3.0.0
aliyun-python-sdk-kms>=2.16.5,<3.0.0
crcmod>=1.7,<2.0.0
langdetect>=1.0.9,<2.0.0
loguru>=0.7.3,<1.0.0
python-docx>=1.1.2,<2.0.0
python-dotenv>=1.1.0,<2.0.0
pymupdf>=1.26.0,<2.0.0
pypdf>=5.5.0,<6.0.0
openpyxl>=3.1.5,<4.0.0
pandas>=2.2.3,<3.0.0
numpy>=2.2.6,<3.0.0
requests>=2.32.3,<3.0.0
tqdm>=4.67.1,<5.0.0
pydantic>=2.11.5,<3.0.0
pydantic-settings>=2.9.1,<3.0.0
python-magic>=0.4.27,<1.0.0
PyYAML>=6.0.2,<7.0.0
Pillow>=11.2.1,<12.0.0
packaging>=24.2,<25.0
beautifulsoup4>=4.13.4,<5.0.0
minio>=7.2.15,<8.0.0
openai>=1.82.0,<2.0.0
jionlp>=1.5.23,<2.0.0
chardet>=5.2.0,<6.0.0
olefile>=0.46
python-pptx>=1.0.2,<2.0.0
tiktoken>=0.9.0,<1.0.0
markitdown>=0.1.1,<1.0.0
xlrd>=2.0.1,<3.0.0
tabulate>=0.9.0,<1.0.0
unstructured>=0.17.2,<1.0.0
markdown>=3.8,<4.0.0
langchain>=0.3.0,<1.0.0
langchain-community>=0.3.0,<1.0.0
langchain-text-splitters>=0.3.0,<1.0.0
ebooklib==0.19
setuptools
transformers==4.53.1
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-timeout>=2.1.0
